Added model sync order

+++ src/index.js
@@ -4,7 +4,7 @@ const ApiRouter = require('./routes/api_routes');
 const bodyParser = require("body-parser");
 const db = require('./config/db_config');
 const cookieParser = require('cookie-parser');
-const { Cart } = require('./models');
+const { syncDbInOrder } = require('./models');
 const app = express();
 
 app.use(bodyParser.json());
@@ -15,17 +15,19 @@ app.use('/api', ApiRouter);
 
 app.listen(serverConfig.PORT, async() => {
     console.log('Server for shop cart is up at port', serverConfig.PORT);
-
-    if (serverConfig.DB_FORCE === true) {
-        await db.sync({ force: true });
-    } else if (serverConfig.DB_ALTER === true) {
-        await db.sync({ alter: true });
-    } else {
-        await db.sync()
+    if (serverConfig.NODE_ENV === 'dev') {
+        if (serverConfig.DB_FORCE === true) {
+            await db.sync({ force: true });
+        } else if (serverConfig.DB_ALTER === true) {
+            await db.sync({ alter: true });
+        } else {
+            await db.sync()
+        }
+    }
+    if (serverConfig.NODE_ENV === 'production') {
+        syncDbInOrder()
     }
 
-    const cart = await Cart.findByPk(1);
-    const products = await cart.getProducts();
-    console.log(products);
+
 
 });
